refactor(scorecard): restructure scorecard configuration, update tests and tooltip types
