+++ workspaces/scorecard/.changeset/stupid-knives-wonder.md
@@ -16,3 +16,22 @@
 - `result.averageScore` → `result.weightedStatusScore`
 - `result.averageWeightedSum` → `result.weightedStatusSum`
 - `result.averageMaxPossible` → `result.weightedStatusMaxPossible`
+
+### Frontend translations (`metric.*`)
+
+Update any scorecard translation overrides that used the old `average*` keys:
+
+- `metric.averageCenterTooltipTotalLabel` → `metric.weightedStatusScoreCenterTooltipTotalLabel`
+- `metric.averageCenterTooltipMaxLabel` → `metric.weightedStatusScoreCenterTooltipMaxLabel`
+- `metric.averageCenterTooltipBreakdownRow_one` / `_other` → `metric.weightedStatusScoreCenterTooltipBreakdownRow_one` / `_other`
+- `metric.averageLegendTooltipEntitiesEach_one` / `_other` → `metric.weightedStatusScoreLegendTooltipEntitiesEach_one` / `_other`
+- `metric.averageLegendTooltipRowTotal` → `metric.weightedStatusScoreLegendTooltipRowTotal`
+
+The `averageLegendTooltip*` keys served the removed side-legend tooltip; the center donut tooltip uses the `weightedStatusScoreCenterTooltip*` keys (including per-status breakdown rows).
+
+### UI `data-testid` values
+
+Update downstream e2e selectors that targeted the weighted-status-score card center label:
+
+- `average-card-center-percent` → `weighted-status-score-card-center-percent`
+- `average-card-center-percent-hit-area` → `weighted-status-score-card-center-percent-hit-area`
