+++ workspaces/scorecard/plugins/scorecard/src/components/AggregatedMetricCards/components/CardTooltip.tsx
@@ -21,7 +21,7 @@ import {
 import Box from '@mui/material/Box';
 import Portal from '@mui/material/Portal';
 import type { PieData } from '../../types';
-import { TooltipPosition } from '../WeightedStatusScoreCard/types';
+import type { TooltipPosition } from '../types';
 
 type CardTooltipProps = {
   tooltipPosition: TooltipPosition;
